+++ crates/libwebrtc-sys/include/bridge.h
@@ -266,6 +266,11 @@ std::unique_ptr<AudioSourceInterface> create_audio_source(
     uint16_t device_index,
     const std::unique_ptr<AudioProcessing>& ap);
 
+// Creates a new `AudioSourceInterface` for Display audio.
+std::unique_ptr<AudioSourceInterface> create_display_audio_source(
+    const AudioDeviceModule& audio_device_module,
+    rust::String device_id);
+
 // Disposes the `AudioSourceInterface` with the provided device ID.
 void dispose_audio_source(const AudioDeviceModule& audio_device_module,
                           rust::String device_id);