+++ crates/native/src/media/source.rs
@@ -38,6 +38,10 @@ pub enum MediaTrackSource<T> {
 #[derive(Clone, Copy, Debug, Default, Eq, Hash, PartialEq)]
 pub struct AudioLevelObserverId(u64);
 
+/// [`AudioDeviceId`] of the default system audio capture device.
+#[cfg(target_os = "windows")]
+pub const SYSTEM_AUDIO_DEVICE_ID: &str = "__DISPLAY_AUDIO__";
+
 /// [`sys::AudioSourceInterface`] wrapper.
 #[derive(AsRef)]
 pub struct AudioSource {